% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_descr.R
\name{clean_descr}
\alias{clean_descr}
\title{Clean Description}
\usage{
clean_descr(x = controle)
}
\arguments{
\item{x}{tibble.}
}
\description{
Recebe um df e retorna um df com a descricao solicitada em um formato 'clean'
}
