% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_plan_contr.R
\name{update_plan_contr}
\alias{update_plan_contr}
\title{Atualiza Plan Controle}
\usage{
update_plan_contr(x = controle2, y = cis2)
}
\arguments{
\item{x}{tibble.}
}
\description{
Recebe dois dfs e retorna um novo df com as informações consolidadas.
}
